+++ packages/custom_lint_builder/lib/src/ignore.dart
@@ -1,4 +1,7 @@
-import 'package:analyzer/error/error.dart' hide LintCode;
+import 'package:analyzer/error/error.dart'
+    hide
+        // ignore: undefined_hidden_name, Needed to support lower analyzer versions
+        LintCode;
 import 'package:collection/collection.dart';
 import 'package:custom_lint_core/custom_lint_core.dart';