+++ packages/custom_lint_visitor/lib/src/node_lint_visitor.g.dart
@@ -9,10 +9,8 @@ part of 'node_lint_visitor.dart';
 // **************************************************************************
 
 /// The AST visitor that runs handlers for nodes from the [_registry].
-@internal
 class LinterVisitor extends GeneralizingAstVisitor<void> {
   /// The AST visitor that runs handlers for nodes from the [_registry].
-  @internal
   LinterVisitor(this._registry);
 
   final NodeLintRegistry _registry;
@@ -86,6 +84,12 @@ class LinterVisitor extends GeneralizingAstVisitor<void> {
   void visitAssignmentExpression(AssignmentExpression node) {
     _runSubscriptions(node, _registry._forAssignmentExpression);
     super.visitAssignmentExpression(node);
+  }
+
+  @override
+  void visitAugmentationImportDirective(AugmentationImportDirective node) {
+    _runSubscriptions(node, _registry._forAugmentationImportDirective);
+    super.visitAugmentationImportDirective(node);
   }
 
   @override
@@ -677,6 +681,12 @@ class LinterVisitor extends GeneralizingAstVisitor<void> {
   }
 
   @override
+  void visitLibraryAugmentationDirective(LibraryAugmentationDirective node) {
+    _runSubscriptions(node, _registry._forLibraryAugmentationDirective);
+    super.visitLibraryAugmentationDirective(node);
+  }
+
+  @override
   void visitLibraryDirective(LibraryDirective node) {
     _runSubscriptions(node, _registry._forLibraryDirective);
     super.visitLibraryDirective(node);
@@ -812,12 +822,6 @@ class LinterVisitor extends GeneralizingAstVisitor<void> {
   void visitNullAssertPattern(NullAssertPattern node) {
     _runSubscriptions(node, _registry._forNullAssertPattern);
     super.visitNullAssertPattern(node);
-  }
-
-  @override
-  void visitNullAwareElement(NullAwareElement node) {
-    _runSubscriptions(node, _registry._forNullAwareElement);
-    super.visitNullAwareElement(node);
   }
 
   @override
@@ -1268,10 +1272,8 @@ class _Subscription<T> {
 }
 
 /// The container to register visitors for separate AST node types.
-@internal
 class NodeLintRegistry {
   /// The container to register visitors for separate AST node types.
-  @internal
   NodeLintRegistry(this._lintRegistry, {required bool enableTiming})
       : _enableTiming = enableTiming;
 
@@ -1342,6 +1344,14 @@ class NodeLintRegistry {
   void addAssignmentExpression(
       String key, void Function(AssignmentExpression node) listener) {
     _forAssignmentExpression
+        .add(_Subscription(listener, _getTimer(key), Zone.current));
+  }
+
+  final List<_Subscription<AugmentationImportDirective>>
+      _forAugmentationImportDirective = [];
+  void addAugmentationImportDirective(
+      String key, void Function(AugmentationImportDirective node) listener) {
+    _forAugmentationImportDirective
         .add(_Subscription(listener, _getTimer(key), Zone.current));
   }
 
@@ -2008,6 +2018,14 @@ class NodeLintRegistry {
         .add(_Subscription(listener, _getTimer(key), Zone.current));
   }
 
+  final List<_Subscription<LibraryAugmentationDirective>>
+      _forLibraryAugmentationDirective = [];
+  void addLibraryAugmentationDirective(
+      String key, void Function(LibraryAugmentationDirective node) listener) {
+    _forLibraryAugmentationDirective
+        .add(_Subscription(listener, _getTimer(key), Zone.current));
+  }
+
   final List<_Subscription<LibraryDirective>> _forLibraryDirective = [];
   void addLibraryDirective(
       String key, void Function(LibraryDirective node) listener) {
@@ -2154,13 +2172,6 @@ class NodeLintRegistry {
   void addNullAssertPattern(
       String key, void Function(NullAssertPattern node) listener) {
     _forNullAssertPattern
-        .add(_Subscription(listener, _getTimer(key), Zone.current));
-  }
-
-  final List<_Subscription<NullAwareElement>> _forNullAwareElement = [];
-  void addNullAwareElement(
-      String key, void Function(NullAwareElement node) listener) {
-    _forNullAwareElement
         .add(_Subscription(listener, _getTimer(key), Zone.current));
   }
 
@@ -2663,10 +2674,8 @@ class NodeLintRegistry {
 class LintRuleNodeRegistry {
   LintRuleNodeRegistry(this.nodeLintRegistry, this.name);
 
-  @internal
   final NodeLintRegistry nodeLintRegistry;
 
-  @internal
   final String name;
 
   @preferInline
@@ -2714,6 +2723,12 @@ class LintRuleNodeRegistry {
   void addAssignmentExpression(
       void Function(AssignmentExpression node) listener) {
     nodeLintRegistry.addAssignmentExpression(name, listener);
+  }
+
+  @preferInline
+  void addAugmentationImportDirective(
+      void Function(AugmentationImportDirective node) listener) {
+    nodeLintRegistry.addAugmentationImportDirective(name, listener);
   }
 
   @preferInline
@@ -3245,6 +3260,12 @@ class LintRuleNodeRegistry {
   }
 
   @preferInline
+  void addLibraryAugmentationDirective(
+      void Function(LibraryAugmentationDirective node) listener) {
+    nodeLintRegistry.addLibraryAugmentationDirective(name, listener);
+  }
+
+  @preferInline
   void addLibraryDirective(void Function(LibraryDirective node) listener) {
     nodeLintRegistry.addLibraryDirective(name, listener);
   }
@@ -3359,11 +3380,6 @@ class LintRuleNodeRegistry {
   @preferInline
   void addNullAssertPattern(void Function(NullAssertPattern node) listener) {
     nodeLintRegistry.addNullAssertPattern(name, listener);
-  }
-
-  @preferInline
-  void addNullAwareElement(void Function(NullAwareElement node) listener) {
-    nodeLintRegistry.addNullAwareElement(name, listener);
   }
 
   @preferInline